requirements.txt
pandas>=2.0.0
numpy>=1.24.0
numpy-financial>=1.0.0
yfinance>=0.2.36
pyyaml>=6.0
openpyxl>=3.1.0
streamlit>=1.32.0
plotly>=5.18.0
matplotlib>=3.7.0
pytest>=7.4.0
rich>=13.0.0
